Fe feature/마이페이지

See merge request s10-bigdata-dist-sub2/S10P22A607!151

+++ frontend/dangmoca-project/src/components/review/SimpleReviewCard.tsx
@@ -3,6 +3,8 @@ import { reviewAPI } from "../../api/reviewAPI";
 import ThumbUp from "../../assets/icons/thumbup.svg?react";
 import { tagMapper } from "../../utils/tag";
 import { useDragScroll } from "../../utils/useDragScroll";
+import { memberAPI } from "../../api/memberAPI";
+import NoNope from "../../assets/icons/nope-no.gif"
 
 // interface Review {
 //   reviewSeq: number;
@@ -30,24 +32,35 @@ const SimpleReviewCard = ({ refreshReviews, ...review }: any) => {
   }
 
   const deletReview = async () => {
-    try {
-      const respronse = await reviewAPI.deleteReview(review.reviewSeq);
-      console.log("리뷰 삭제 성공!", respronse.data);
+    const myResponse = await memberAPI.getMyInfo();
+    if (myResponse.data.result.memberSeq === review.memberSeq){
+      try {
+        const respronse = await reviewAPI.deleteReview(review.reviewSeq);
+        console.log("리뷰 삭제 성공!", respronse.data);
+        Swal.fire({
+          title: "삭제 성공!",
+          icon: "success",
+        }).then(() => refreshReviews());
+      } catch (error) {
+        Swal.fire({
+          title: "리뷰 삭제에 실패했습니다.",
+          icon: "error",
+        });
+        console.log(error);
+      }
+    } else {
       Swal.fire({
-        title: "삭제 성공!",
-        icon: "success",
-      }).then(() => refreshReviews());
-    } catch (error) {
-      Swal.fire({
-        title: "권한을 확인해 주세요",
+        title: "권한을 확인해 주세요.",
         icon: "error",
       });
-      console.log(error);
     }
   };
 
   const modifyReview = () => {
-    console.log("더미");
+    Swal.fire({
+      title: "공정한 리뷰 문화!",
+      icon: "error",
+    });
   };
 
   return (
@@ -82,12 +95,13 @@ const SimpleReviewCard = ({ refreshReviews, ...review }: any) => {
       </div>
       <div className="flex justify-between py-2">
         <button
-          onClick={
-            () =>
-              Swal.fire({
-                title: "공정한 리뷰 문화!",
-                icon: "error",
-              })
+          onClick={() =>
+            Swal.fire({
+              title: "공정한 리뷰 문화!",
+              imageUrl: `${NoNope}`,
+              imageHeight: 300,
+              imageWidth: 300
+            })
           }
         >
           <ThumbUp id="svgIcon" className="mr-2" />
